Tests/LibWeb/Text/expected/HTML/HTMLOrSVGElement-nonce.txt

generic IDL and attribute interaction
nonce: "" attribute: null;
nonce: "123" attribute: null;
nonce: "456" attribute: 456;
nonce: "null" attribute: 456;
insertion
nonce: "foo" attribute: foo;
nonce: "foo" attribute: foo;
cloning
nonce: "bar" attribute: bar;
nonce: "bar" attribute: bar;
